requirements.txt
requests
enum34
pyyaml
spotipy==2.9.0
pymongo
pyOpenSSL
autobahn[twisted,serialization]
PyQt5